+++ sdk/sourcefiles/native/internal/ANNativeMediatedAdController.h
@@ -16,9 +16,8 @@
 #import "ANMediatedAd.h"
 #import "ANAdProtocol.h"
 #import "ANNativeAdFetcher.h"
-
-
-
+#import "ANTargetingParameters.h"
+#import "ANNativeCustomAdapter.h"
 @protocol ANNativeMediationAdControllerDelegate;
 
 
@@ -26,9 +25,48 @@
 
 @interface ANNativeMediatedAdController : NSObject
 
-// Designated initializer
+// variables for measuring latency.
+@property (nonatomic, readwrite, assign) NSTimeInterval latencyStart;
+@property (nonatomic, readwrite, assign) NSTimeInterval latencyStop;
+
+@property (nonatomic, readwrite, assign) BOOL hasSucceeded;
+@property (nonatomic, readwrite, assign) BOOL hasFailed;
+@property (nonatomic, readwrite, assign) BOOL timeoutCanceled;
+
+
+@property (nonatomic, readwrite, weak)  ANNativeAdFetcher  *adFetcher;
+@property (nonatomic, readwrite, weak)  id<ANNativeAdFetcherDelegate>     adRequestDelegate;
+
+// Designated initializer for CSM Mediation
 + (instancetype)initMediatedAd:(ANMediatedAd *)mediatedAd
                    withFetcher: (ANNativeAdFetcher *)adFetcher
              adRequestDelegate:(id<ANNativeAdRequestProtocol>)adRequestDelegate;
+
+
+
+// Adapter helper methods
+- (ANTargetingParameters *)targetingParameters;
+- (NSString *)createResponseURLRequest:(NSString *)baseString reason:(int)reasonCode;
+- (void)didReceiveAd:(id)adObject;
+- (BOOL)checkIfMediationHasResponded;
+- (void)setAdapter:(id<ANNativeCustomAdapter>)adapter;
+- (void)clearAdapter;
+- (void)didFailToReceiveAd:(ANAdResponseCode)errorCode;
+
+
+// Adapter Error Handling
+- (void)finish:(ANAdResponseCode)errorCode withAdObject:(id)adObject;
+- (void)handleInstantiationFailure:(NSString *)className errorCode:(ANAdResponseCode)errorCode errorInfo:(NSString *)errorInfo;
+
+
+// Adapter Latency Measurement
+- (void)markLatencyStart;
+- (void)markLatencyStop;
+- (NSTimeInterval)getLatency;
+
+// Adapter Timeout handler
+- (void)startTimeout;
+- (void)cancelTimeout;
+
 @end